[refactoring] TestShell 클래스의 일부 역할을 CommandParser 클래스로 이관
