+++ tests/testapp_test/test_testshell.py
@@ -1,10 +1,10 @@
 from unittest import TestCase
 from unittest.mock import patch
 
+from testapp.command import Read, Write, FullRead, FullWrite
 from testapp.test_app1 import TestApp1
 from testapp.test_app2 import TestApp2
 from testapp.testshell import TestShell, EXECUTE_INVALID, EXECUTE_VALID_WO_ARGS, EXECUTE_VALID_WITH_ARGS
-from testapp.command import Read, Write, FullRead, FullWrite
 
 
 class TestTestShell(TestCase):
@@ -18,117 +18,6 @@ class TestTestShell(TestCase):
         self.ts = TestShell()
         TestTestShell.test_counter += 1
         print(f"\nRunning test #{TestTestShell.test_counter}")
-
-    def test_valid_cmd_true(self):
-        cmd_list = [
-            "write 3 0xAAAABBBB",
-            "read 3",
-            "help",
-            "fullwrite 0xABCDFFFF",
-            "fullread",
-            "testapp1",
-            "testapp2",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertTrue(ret)
-
-    def test_valid_cmd_write_false(self):
-
-        cmd_list = [
-            "Write 3 0xAAAABBBB",
-            "Write 3 0xAAAABBBB",
-            "write 0xAAAABBBB",
-            "write 3",
-            "write",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertFalse(ret)
-
-    def test_valid_cmd_read_false(self):
-
-        cmd_list = [
-            "read 3 0xAAAABBBB",
-            "Read 3",
-            "read 0xAAAABBBB",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertFalse(ret)
-
-    def test_valid_cmd_exit_false(self):
-
-        cmd_list = [
-            "Exit",
-            "exit 1",
-            "exit 0xABCDFFFF",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertFalse(ret)
-
-    def test_valid_cmd_help_false(self):
-
-        cmd_list = [
-            "HELP",
-            "hELP",
-            "help 0xABCDFFFF",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertFalse(ret)
-
-    def test_valid_cmd_fullwrite_false(self):
-
-        cmd_list = [
-            "Fullwrite 0xABCDFFFF",
-            "FULLWRITE 0xABCDFFFF",
-            "FullWrite 0xABCDFFFF",
-            "fullwrite ABCDFFFF",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertFalse(ret)
-
-    def test_valid_cmd_fullread_false(self):
-
-        cmd_list = [
-            "Fullread",
-            "FULLREAD",
-            "FullRead",
-        ]
-        for idx, cmd in enumerate(cmd_list):
-            with self.subTest(idx=idx, cmd=cmd):
-                print(f"{idx + 1}. {cmd}")
-                ret = self.ts.validate_command(cmd)
-                self.assertFalse(ret)
-
-    def test_parse_args(self):
-        cmd_dict = {
-            "write 3 0xAAAABBBB": ("write", [3, 0xAAAABBBB]),
-            "read 3": ("read", [3]),
-            "help": ("help", None),
-            "fullwrite 0xABCDFFFF": ("fullwrite", [0xABCDFFFF]),
-            "fullread": ("fullread", None),
-        }
-        for idx, (key, val) in enumerate(cmd_dict.items()):
-            with self.subTest(idx=idx, key=key, val=val):
-                print(f"{idx + 1}. {key},{val}")
-                ret = self.ts.parse_args(key)
-                self.assertEqual(ret, val)
 
     @patch.object(TestApp2, "run", return_value=True)
     @patch.object(TestApp1, "run", return_value=True)
